I moved report folder out of campaign folder, I created a enum for type in emails report and other fixes.
