+++ src/main/java/com/ecwid/maleorang/method/v3_0/reports/CampaignBounceReport.kt
@@ -1,4 +1,4 @@
-package com.ecwid.maleorang.method.v3_0.campaign.report
+package com.ecwid.maleorang.method.v3_0.reports
 
 import com.ecwid.maleorang.MailchimpObject
 import com.ecwid.maleorang.annotation.Field
@@ -8,17 +8,16 @@ import java.util.*
  * Created by: Manuel Lara <lararojas.mr@gmail.com>
  */
 
-class CampaignFacebookLikesReport : MailchimpObject() {
+class CampaignBounceReport : MailchimpObject() {
     @JvmField
     @Field
-    var recipient_likes: Int? = null
+    var hard_bounces: Int? = null
+
+    @JvmField
+    @Field
+    var soft_bounces: Int? = null
 	
     @JvmField
     @Field
-    var unique_likes: Int? = null	
-	
-    @JvmField
-    @Field
-    var facebook_likes: Int? = null
-	
+    var syntax_errors: Int? = null
 }
