+++ src/main/java/com/ecwid/maleorang/method/v3_0/reports/CampaignDeliveryStatusReport.kt
@@ -1,4 +1,4 @@
-package com.ecwid.maleorang.method.v3_0.campaign.report
+package com.ecwid.maleorang.method.v3_0.reports
 
 import com.ecwid.maleorang.MailchimpObject
 import com.ecwid.maleorang.annotation.Field
