+++ src/main/java/com/ecwid/maleorang/method/v3_0/reports/CampaignEcommerceReport.kt
@@ -1,4 +1,4 @@
-package com.ecwid.maleorang.method.v3_0.campaign.report
+package com.ecwid.maleorang.method.v3_0.reports
 
 import com.ecwid.maleorang.MailchimpObject
 import com.ecwid.maleorang.annotation.Field
@@ -8,16 +8,17 @@ import java.util.*
  * Created by: Manuel Lara <lararojas.mr@gmail.com>
  */
 
-class CampaignBounceReport : MailchimpObject() {
+class CampaignEcommerceReport : MailchimpObject() {
     @JvmField
     @Field
-    var hard_bounces: Int? = null
+    var total_orders: Int? = null
 
     @JvmField
     @Field
-    var soft_bounces: Int? = null
+    var total_spent: Double? = null
 	
     @JvmField
     @Field
-    var syntax_errors: Int? = null
+    var total_revenue: Double? = null	
+	
 }
