+++ src/main/java/com/ecwid/maleorang/method/v3_0/reports/CampaignFacebookLikesReport.kt
@@ -1,4 +1,4 @@
-package com.ecwid.maleorang.method.v3_0.campaign.report
+package com.ecwid.maleorang.method.v3_0.reports
 
 import com.ecwid.maleorang.MailchimpObject
 import com.ecwid.maleorang.annotation.Field
@@ -8,17 +8,17 @@ import java.util.*
  * Created by: Manuel Lara <lararojas.mr@gmail.com>
  */
 
-class CampaignEcommerceReport : MailchimpObject() {
+class CampaignFacebookLikesReport : MailchimpObject() {
     @JvmField
     @Field
-    var total_orders: Int? = null
-
-    @JvmField
-    @Field
-    var total_spent: Double? = null
+    var recipient_likes: Int? = null
 	
     @JvmField
     @Field
-    var total_revenue: Double? = null	
+    var unique_likes: Int? = null	
+	
+    @JvmField
+    @Field
+    var facebook_likes: Int? = null
 	
 }
