+++ src/main/java/com/ecwid/maleorang/method/v3_0/reports/CampaignListStatsReport.kt
@@ -1,4 +1,4 @@
-package com.ecwid.maleorang.method.v3_0.campaign.report
+package com.ecwid.maleorang.method.v3_0.reports
 
 import com.ecwid.maleorang.MailchimpObject
 import com.ecwid.maleorang.annotation.Field
@@ -8,21 +8,21 @@ import java.util.*
  * Created by: Manuel Lara <lararojas.mr@gmail.com>
  */
 
-class CampaignTimeseriesReport : MailchimpObject() {
+class CampaignListStatsReport : MailchimpObject() {
     @JvmField
     @Field
-    var timestamp: Date? = null
+    var sub_rate: Double? = null
 
     @JvmField
     @Field
-    var emails_sent: Int? = null
+    var unsub_rate: Double? = null
 	
     @JvmField
     @Field
-    var unique_opens: Int? = null
-		
+    var open_rate: Double? = null	
+	
     @JvmField
     @Field
-    var recipients_clicks: Int? = null
+    var click_rate: Double? = null
 	
 }
