+++ src/main/java/com/ecwid/maleorang/method/v3_0/reports/email/EmailReportActivityDetails.kt
@@ -1,4 +1,4 @@
-package com.ecwid.maleorang.method.v3_0.campaign.report
+package com.ecwid.maleorang.method.v3_0.reports.email
 
 import com.ecwid.maleorang.MailchimpObject
 import com.ecwid.maleorang.annotation.Field
@@ -8,25 +8,25 @@ import java.util.*
  * Created by: Manuel Lara <lararojas.mr@gmail.com>
  */
 
-class CampaignClicksReport : MailchimpObject() {
+class EmailReportActivityDetails : MailchimpObject() {
     @JvmField
     @Field
-    var clicks_total: Int? = null
+    var action: String? = null
 
     @JvmField
     @Field
-    var unique_clicks: Int? = null
+    var type: TypeEmailActivity? = null
 	
     @JvmField
     @Field
-    var unique_subscriber_clicks: Int? = null	
+    var timestamp: Date? = null
 	
     @JvmField
     @Field
-    var click_rate: Double? = null
+    var url: String? = null
 		
     @JvmField
     @Field
-    var last_click: Date? = null
+    var ip: String? = null
 	
 }
