+++ src/main/java/com/ecwid/maleorang/method/v3_0/reports/email/TypeEmailActivity.kt
@@ -1,0 +1,11 @@
+package com.ecwid.maleorang.method.v3_0.reports.email
+
+/**
+ * Created by larar on 15/10/2016.
+ */
+
+enum class TypeEmailActivity(val type: String) {
+    BOUNCE("bounce"),
+    CLICK("click"),
+    OPEN("open")
+}
